functions/converter/requirements.txt
flask==3.0.2
google-cloud-storage==2.14.0 